fix: Merge conflicts [DHIS2-14206]

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/visualization/AxisV2.java
@@ -72,12 +72,12 @@ public class AxisV2 implements Serializable
     @JsonProperty
     @JacksonXmlProperty( namespace = DXF_2_0 )
     @PropertyRange( min = -MAX_VALUE )
-    private Integer maxValue;
+    private Double maxValue;
 
     @JsonProperty
     @JacksonXmlProperty( namespace = DXF_2_0 )
     @PropertyRange( min = -MAX_VALUE )
-    private Integer minValue;
+    private Double minValue;
 
     @JsonProperty
     @JacksonXmlProperty( namespace = DXF_2_0 )
